Update Seminar Sub-page Sidebar Look
The sidebars are now pills. However, both links will go to the
"upcoming" tab of the calendar page. From my research, we won't be able
to link to a specific tab from this page without using javascript - so
maybe we should consider getting rid of the sidebar on the individual seminar pages altogether...