Updated Seminar TBD Behavior
For the front page:
- If there is no title or speakers assigned, 'DUB Seminar: Speaker TBD'
will be displayed.
- If there is no title, but speaker(s) are assigned, 'DUB Seminar:
(speaker name(s))' will be displayed.
- If location is TBD, it will not be displayed.

For the calendar page:
- If there is no title or speakers assigned, 'DUB Seminar: Speaker TBD'
will be displayed as a title with no link.
- If there is no title but speaker(s) are assigned, no title will be
displayed, but the speaker(s) will be displayed.
- If location is TBD, it will not be displayed

For the seminar page:
- If there is no title or speakers assigned, this page will not be
linked to.
- Any field that is TBD (except for date, time) will not be displayed.

I also made a CSS tweak to align headers with normal text on the
calendar page, so things aren't misaligned when there is no title but
there are speakers for a seminar entry.