Subscription link
#257
`webcal://` works for outlook & apple calendar app, but not Google. The
Google subscribe link no longer seems to work. This got me closer, but
I still got an error:
http://stackoverflow.com/questions/12563178/link-to-add-ics-to-google-ca
lendar-stopped-working

So this PR is the catch-all of a link to the file and the URL in text.
People can copy either the string or the link address into their
calendar app.